Add interface for other components to build an instance of an LLM
Currently, we have a few places that instantiate an LLM class
(vectorsearch and agent).

The current implementation ends up looking at
a constant to find a mapping from a name to a class name, and then uses
const_get on Langchain::LLM::Base to get instantiate it. They do use a
method from Base to make sure it's valid first. But even with that, the
component using an LLM ends up pretty coupled to the LLM data structure
for tracking what it supports.

Since we are using a name, and building an instance, it made me think
factory pattern. And thinking factory pattern, reminded me of
factory_bot. That inspired this interface:

    Langchain::LLM.build(:openai, api_key: "123")

I almost left it there, but something that has been bugging me is how to
allow third party LLMs that aren't part of the repo. The current way
does requires it to be in that hash of names to class names. I got to
thinking, maybe we can support passing in an LLM object directly? In
that case, we don't need to do anything to look it up.

This actually lines up with how langchain.py is simplemented:

    llm = OpenAI(temperature=0.9)
    prompt = PromptTemplate(
        input_variables=["product"],
        template="What is a good name for a company that makes {product}?",
    )
    chain = LLMChain(llm=llm, prompt=prompt)

I ended up adding these to the LLM module directly, and moved some logic
from LLM::Base, which left that class solely responsible for a base
class, instead of also being responsible for validating and knowing what
LLMs are supported.