Update Pinecone Ruby Client to 1.x
This PR updates Pinecone Ruby Client to 1.x to use Pinecone V2 API.

Pinecone 1.1 has already been released:
https://github.com/ScotterC/pinecone/blob/main/CHANGELOG.md

So this change sets the development dependency to `"~> 1.0"` to enable testing within the 1.x series.
As noted in issue #662, users are likely specifying `"~> 1.0"` in their Gemfile,
so this version constraint is chosen to allow compatibility with minor updates.

Closes #662.