Emit deprecation warning for `default_dimension` method (#990)
Based on https://github.com/patterns-ai-core/langchainrb/commit/16dfeb2,
it seems that `default_dimension` is being kept for backward compatibility.

Emitting a warning when `default_dimension` is used would help encourage users to
switch to the regular `default_dimensions` method, and would be a suitable way to
avoid breaking change.