Refine warning behavior for unsupported options (#983)
## Summary

Currently, the warning for unsupported options is shown regardless of whether an explicit value is provided.

This PR updates the behavior so that the warning is shown only when the unsupported options is explicitly set by the user.

Tests have been updated to reflect usage without warnings, which is what users would expect.

## Additional Information

Note that `parallel_tool_calls` option in `Langchain::Assistant#initialize` defaults to `true`:
https://github.com/patterns-ai-core/langchainrb/blob/0.19.5/lib/langchain/assistant.rb#L45

However, whether this option is supported depends on the underlying LLM.
Therefore, for LLMs that do not support `parallel_tool_calls`, the option should effectively not be `true`.

Since this would involve changing the default behavior of `parallel_tool_calls`,
it is considered out of scope for this PR and should be addressed separately.