Rename Langchain.rb to LangChain.rb
In the LLM field, the name LangChain is widely recognized, and the project was
likely originally intended to be called LangChain.rb as a product:
https://github.com/patterns-ai-core/langchainrb/commit/ab9d9889c2183d99a90bee1347fe1fed0119e4c1

It appears that the current use of the name `Langchain` is due to implementation constraints.
This PR updates the behavior so that Zeitwerk also recognizes the `LangChain` module name for the `langchain.rb` file.

Both the existing `Langchain` and the new `LangChain` module names will work.
However, the `Langchain` constant is now marked as deprecated, and a warning will be shown when it is used:

```console
$ bundle exec ruby -Ilib -rrails -rlangchainrb -e 'Langchain::Errors'
-e:1: warning: `LangChain` is deprecated. Use `LangChain` instead.
```

This change makes it possible to use the name `LangChain`, which is the standard name in the AI ecosystem.
Because `LangChain` is a proper noun, renaming to this widely recognized and commonly used name follows
the principle of least surprise for users.